requirements.txt
openai>=1.14.0
pinecone>=3.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.4.0
streamlit>=1.33.0
tqdm>=4.66.0
pdfplumber>=0.10.0
python-docx>=1.1.0
PyPDF2>=3.0.0
gspread>=5.12.0
google-auth>=2.28.0
requests>=2.31.0